mcp-gw/pyproject.toml
[project]
name = "mcp-gw"
version = "0.1.0"
description = "MCP server with mock tool implementations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.43.0",
]

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-anyio>=0.0.0",
]
